Clean up + Merge composite actions (setup/build/test) into a single workflow
Making them reusable workflows are not possible because they would
not be callable as a single job step (which was what composite actions
are for). But the steps in these actions are so tiny and problem-specific
that making them standalone is hard to maintain anyway, so a single,
moderate-sized workflow is acceptable.