_get_arguments_info mustn't treat all error code as equal
As `arg_pos` is being incremented, eventually `CUDA_ERROR_INVALID_VALUE`
is expected when `arg_pos` exceeds the number of kernel arguments.

But other errors like CUDA_ERROR_INVALID_CONTEXT are possible, and they
should be treated as errors.

Failure to do that was behind the unintuitive behavior I reported in the PR.