Implement Kernel.num_arguments, and Kernel.arguments_info (#612)
* Implement Kernel.num_arguments, and Kernel.arguments_info

* Factor out common logic between num_arguments and arguments_info properties

Also parametrize test to check arguments_info to check with all int,
and all short arguments.

* Use namedtuple turn to pack offset/size for individual kernel argument

* Add feature entry to 0.3.0 release notes

* Move ParamInfo to class context

Used modern namedtuple instance constructor.

* _get_arguments_info mustn't treat all error code as equal

As `arg_pos` is being incremented, eventually `CUDA_ERROR_INVALID_VALUE`
is expected when `arg_pos` exceeds the number of kernel arguments.

But other errors like CUDA_ERROR_INVALID_CONTEXT are possible, and they
should be treated as errors.

Failure to do that was behind the unintuitive behavior I reported in the PR.

* Add skip if testing with compute sanitizer, skip if cuda version < 12

Also add a test that if cuda is not initialized, an error is raised when
checking for the num_arguments.

* Implement suggestions from @leofang

1. Narrow fixture scope
2. Rename fixture to cuda12_prerequisite_check that provide a boolean
   rather than a pair of versions

* Use ParamInfo instead of NamedTuple is annotation

This required moving ParamInfo definition from class scope
to module scope, since referencing Kernel.ParamInfo from
annotations of methods of the Kernel class results in error
that Kernel class does not yet exist.